requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
idna==2.8
ipython-genutils==0.2.0
joblib==0.13.2
jsonschema==3.0.1
jupyter-core==4.5.0
kiwisolver==1.1.0
matplotlib==3.1.0
nbformat==4.4.0
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
plotly==3.10.0
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
retrying==1.3.3
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
traitlets==4.3.2
urllib3==1.25.3
